% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{M_regression_GS}
\alias{M_regression_GS}
\title{Initial values for image-on-scalar regression}
\usage{
M_regression_GS(
  data,
  init,
  region_idx_cpp,
  kernel,
  n_mcmc,
  display_progress = TRUE
)
}
\arguments{
\item{data}{an R list object that contains the data}

\item{init}{R list object with initial values}

\item{region_idx_cpp}{a list of region idx for all regions starting from 0}

\item{kernel}{a list object containing Q and D}

\item{n_mcmc}{number of iterations}

\item{display_progress}{bool, 1 = display progress bar}
}
\description{
Initial values for image-on-scalar regression
}
